requirements.txt
cffi==1.13.2
cryptography==2.8
lxml==4.5.0
pycparser==2.19
pyOpenSSL==19.1.0
six==1.14.0
urllib3==1.25.8
xmltodict==0.12.0
selinium==3.141.0
beautifulsoup4==4.9.0